ai-service/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
pydantic==2.10.4
python-dotenv==1.0.1
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
networkx==3.2.1
sentence-transformers==2.7.0
psycopg2-binary==2.9.9

fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
psycopg[binary]>=3.1.18
numpy>=1.26.0
scikit-learn>=1.4.0
